+++ src/Microsoft.AspNetCore.Server.IISIntegration/Server/IO/IISWebSocketReadOperation.cs
@@ -1,10 +1,18 @@
 using System;
 using System.Buffers;
+using System.Runtime.InteropServices;
 
 namespace Microsoft.AspNetCore.Server.IISIntegration
 {
-    internal class IISAsyncReadOperation : IISAsyncIOOperation
+    internal class WebSocketReadOperation : AsyncIOOperation
     {
+        private readonly GCHandle _thisHandle;
+
+        public WebSocketReadOperation()
+        {
+            _thisHandle = GCHandle.Alloc(this);
+        }
+
         public void Initialize(IntPtr requestHandler, Memory<byte> memory)
         {
             _requestHandler = requestHandler;
@@ -20,12 +28,15 @@ namespace Microsoft.AspNetCore.Server.IISIntegration
         public override unsafe bool InvokeOperation()
         {
             _inputHandle = _memory.Pin();
-            var hr = NativeMethods.HttpReadRequestBytes(
+
+            var hr = NativeMethods.HttpWebsocketsReadBytes(
                 _requestHandler,
                 (byte*)_inputHandle.Pointer,
                 _memory.Length,
+                ReadCallback,
+                (IntPtr)_thisHandle,
                 out var dwReceivedBytes,
-                out bool fCompletionExpected);
+                out var fCompletionExpected);
 
             if (!fCompletionExpected)
             {
@@ -37,6 +48,11 @@ namespace Microsoft.AspNetCore.Server.IISIntegration
             return false;
         }
 
+        public override void NotifyOperationCompletion(int hr, int bytes)
+        {
+            _inputHandle.Dispose();
+        }
+
         public override void ResetOperation()
         {
             _memory = default;
@@ -45,9 +61,17 @@ namespace Microsoft.AspNetCore.Server.IISIntegration
             _requestHandler = default;
         }
 
-        public override void NotifyOperationCompletion(int hr, int bytes)
+        public static readonly NativeMethods.PFN_WEBSOCKET_ASYNC_COMPLETION ReadCallback = (pHttpContext, pCompletionInfo, pvCompletionContext) =>
         {
-            _inputHandle.Dispose();
-        }
+            var context = (WebSocketReadOperation)GCHandle.FromIntPtr(pvCompletionContext).Target;
+
+            NativeMethods.HttpGetCompletionInfo(pCompletionInfo, out int cbBytes, out int hr);
+
+            var continuation = context.NotifyCompletion(hr, cbBytes);
+
+            continuation.Invoke();
+
+            return NativeMethods.REQUEST_NOTIFICATION_STATUS.RQ_NOTIFICATION_PENDING;
+        };
     }
 }
